Relocate str_diagnostic function from tests to theano.compile.debugmode
